Fixed get log bug
Got one bug report in which user is getting an error when the extension gets the log and tries to parse it as json. The log the user sent was valid json so I believe the extension is recieving an object instead of text like it does on my machine. I'm not sure this is the issue, but if it is this patch should fix it.